Add support for opening URIs in existing instance
Spotify exposes a "org.mpris.MediaPlayer2.Player.OpenUri" dbus method,
that can be used to remotely control an already running Spotify
instance.

We can use this dbus interface to avoid opening multiple Spotify
instances when opening Spotify URLs.

Passing a Spotify URI in the "spotify:<type>:<UUID>" format will make
Spotify end the current song and start playing the given song
immediately.

An example to test this is to run:
xdg-open 'spotify:track:5GJaxfibCRCQXDVPgHJv0s'

This is also the format given by Spotify Web when clicking on "Open in
Desktop App", thus this should work by default everywhere.

If the dbus call fails, we fallback gracefully to the old behaviour,
i.e. opening a new Spotify instance. In theory this should not happen
though, as we check if Spotify is running before making that OpenUri
dbus call.
Obviously the behaviour is also not changed, if there is currently no
Spotify instance running.