Add support for open.spotify.com URLs
Strictly speaking this is not really needed, because usually Spotify
can only pass links with either "spotify:" or "spotify://" scheme.

However with this, it is possible to configure a proxy default browser
such that Spotify links are opened directly in the Spotify desktop
application without first having to confirm "Open in app" in the
browser.